Move package_command into it's own controller
Makes SourceController smaller and both controllers more readable.
