+++ src/api/.rubocop_todo.yml
@@ -85,6 +85,7 @@ Metrics/ClassLength:
     - 'app/controllers/request_controller.rb'
     - 'app/controllers/search_controller.rb'
     - 'app/controllers/source_controller.rb'
+    - 'app/controllers/source_package_command_controller.rb'
     - 'app/controllers/webui/comments_controller.rb'
     - 'app/controllers/webui/kiwi/images_controller.rb'
     - 'app/controllers/webui/package_controller.rb'
@@ -180,6 +181,7 @@ Metrics/CyclomaticComplexity:
     - 'app/controllers/request_controller.rb'
     - 'app/controllers/search_controller.rb'
     - 'app/controllers/source_controller.rb'
+    - 'app/controllers/source_package_command_controller.rb'
     - 'app/controllers/source_project_controller.rb'
     - 'app/controllers/source_project_meta_controller.rb'
     - 'app/controllers/trigger_controller.rb'
@@ -302,6 +304,7 @@ Metrics/PerceivedComplexity:
     - 'app/controllers/request_controller.rb'
     - 'app/controllers/search_controller.rb'
     - 'app/controllers/source_controller.rb'
+    - 'app/controllers/source_package_command_controller.rb'
     - 'app/controllers/source_project_controller.rb'
     - 'app/controllers/source_project_meta_controller.rb'
     - 'app/controllers/webui/monitor_controller.rb'
